Separate manifest for dev site. Moved help dialog comp into own file. Updated readme. check for leading 0 input
