+++ public/index.html
@@ -19,7 +19,16 @@
   <link rel="apple-touch-icon" sizes="180x180" href="./resources/apple-touch-icon.png">
   <link rel="icon" type="image/png" sizes="32x32" href="./resources/favicon-32x32.png">
   <link rel="icon" type="image/png" sizes="16x16" href="./resources/favicon-16x16.png">
-  <link rel="manifest" href="./resources/manifest.json">
+  <script>
+    if(window.location.hostname === 'debug-ezwhatsapp.vercel.app'){
+      console.log('Running test version');
+      document.write('<link rel="manifest" href="./resources/manifest-test.json">');
+    }
+    else{
+      console.log('prod');
+      document.write('<link rel="manifest" href="./resources/manifest.json">');
+    }
+  </script>
   <link rel="mask-icon" href="./resources/safari-pinned-tab.svg" color="#5bbad5">
   <link rel="shortcut icon" href="./resources/favicon.ico">
   <meta name="msapplication-TileColor" content="#da532c">
