+++ public/resources/manifest-test.json
@@ -1,0 +1,21 @@
+{
+    "name": "TEST-Easy WhatsApp",
+    "short_name": "TEST EZ WA",
+    "description": "Easy WA debug version",
+    "icons": [
+        {
+            "src": "android-chrome-192x192.png",
+            "sizes": "192x192",
+            "type": "image/png"
+        },
+        {
+            "src": "android-chrome-512x512.png",
+            "sizes": "512x512",
+            "type": "image/png"
+        }
+    ],
+    "start_url": "/",
+    "theme_color": "#ffA500",
+    "background_color": "#ffffff",
+    "display": "standalone"
+}
