+++ src/index.js
@@ -21,9 +21,6 @@ import {
   ListItemIcon,
   ListItemText,
 } from '@mui/material';
-// import WhatsAppIcon from '@mui/icons-material/WhatsApp';
-// import AndroidIcon from '@mui/icons-material/Android';
-// import AppleIcon from '@mui/icons-material/Apple';
 import {
   WhatsApp as WhatsAppIcon,
   Android as AndroidIcon,
@@ -32,58 +29,13 @@ import {
 // import callCodes from 'country-calling-code';
 // import { searchCallCode } from './countrycode.js';
 import Copyright from './footer.js';
+import HelpDialog from './helpDialog.js';
 import Disclaimer from './disclaimer.js';
 import * as serviceWorkerRegistration from './serviceWorkerRegistration';
 
 
 
 const theme = createTheme();
-
-function HelpDialog(props) {
-  const {open, onClose} = props;
-
-  const handleCloseHelper = () => {
-    onClose(false);
-  };
-
-  return(
-    <Dialog open={open} onClose={handleCloseHelper}>
-      <DialogTitle>Easy WhatsApp</DialogTitle>
-      <DialogContent>
-        <Typography gutterBottom sx={{mb:1}}>
-          Use this app to easily start a WhatsApp conversation without the need to save recipient phone number into your contacts.
-        </Typography>
-        <Typography gutterBottom sx={{mb:1}}>
-          Easily access this app by installing it to your mobile device's home screen:
-        </Typography>
-        <Typography>
-        <Box sx={{ width: '100%'}}>
-          <List>
-            <ListItem>
-              <ListItemButton href='https://bit.ly/3Al9IW1'>
-                <ListItemIcon>
-                  <AndroidIcon/>
-                </ListItemIcon>
-                <ListItemText primary="Browsers on Android" />
-              </ListItemButton>
-            </ListItem>
-            <ListItem>
-              <ListItemButton href='https://apple.co/3Po99io'>
-                <ListItemIcon>
-                  <AppleIcon/>
-                </ListItemIcon>
-                <ListItemText primary="Safari on iOS" />
-              </ListItemButton>
-            </ListItem>
-
-          </List>
-        </Box>
-        </Typography>
-      </DialogContent>
-    </Dialog>
-
-  )
-}
 
 function NumberForm() {
   const [openHelper, setOpenHelper] = useState(false);
@@ -116,8 +68,15 @@ function NumberForm() {
       })
     }
 
+    // Check for leading +
     if(whatsappNum.includes('+')){
       whatsappNum = whatsappNum.substr(1);
+    }
+
+    // IDN Only! If input has leading 0, remove 0 and prepend 62
+    // e.g. 08121234123 -> 628121234123
+    if(whatsappNum.substr(0,1) === '0'){
+      whatsappNum = '62' + whatsappNum.substr(1);
     }
 
     if(!errFlag){
@@ -191,11 +150,11 @@ function NumberForm() {
                 margin='normal'
                 fullWidth
                 id='number'
-                label='Enter Number Here (include country code)'
+                label='Enter Number Here'
                 name='number'
                 placeholder='example: 62812312341234'
                 error={formError}
-                helperText={formError ? errorMsg.value : ''}
+                helperText={formError ? errorMsg.value : 'Please include country code. Indonesian numbers can start with leading 0 (e.g 0812...)'}
                 autoFocus
               />
               <Button
